refactor: move stats batch helpers to stats/api/stats
