+++ apps/web/src/comment/hooks/useActivity.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { useAuth } from '@/shared/hooks/useAuth';
-import { fetchActivityCountsFromSupabase } from '@/shared/api/supabaseReads';
+import { fetchActivityCountsFromSupabase } from '@/stats/api/stats';
 
 // 메인 훅
 export const useActivity = (fromUserId: string, fromDaysAgo: number) => {
